# Improve assertions and enable `TENZIR_ASSERT` in release builds