--- a/libtenzir/builtins/aspects/fields.cpp
+++ b/libtenzir/builtins/aspects/fields.cpp
@@ -82,8 +82,8 @@ auto traverse(type t) -> generator<schema_context> {
     result.type.category = "container";
   else
     result.type.category = "atomic";
-  TENZIR_ASSERT_CHEAP(not caf::holds_alternative<list_type>(t));
-  TENZIR_ASSERT_CHEAP(not caf::holds_alternative<map_type>(t));
+  TENZIR_ASSERT(not caf::holds_alternative<list_type>(t));
+  TENZIR_ASSERT(not caf::holds_alternative<map_type>(t));
   if (const auto* record = caf::get_if<record_type>(&t)) {
     auto i = size_t{0};
     for (const auto& field : record->fields()) {