--- a/libtenzir/builtins/formats/zeek_tsv.cpp
+++ b/libtenzir/builtins/formats/zeek_tsv.cpp
@@ -369,11 +369,11 @@ struct zeek_printer {
     }
 
     auto operator()(view<pattern>) noexcept -> bool {
-      die("unreachable");
+      TENZIR_UNREACHABLE();
     }
 
     auto operator()(view<map>) noexcept -> bool {
-      die("unreachable");
+      TENZIR_UNREACHABLE();
     }
 
     auto operator()(view<std::string> x) noexcept -> bool {