--- a/libtenzir/include/tenzir/arrow_table_slice.hpp
+++ b/libtenzir/include/tenzir/arrow_table_slice.hpp
@@ -140,7 +140,7 @@ template <concrete_type Type>
 view<type_to_data_t<Type>>
 value_at([[maybe_unused]] const Type& type,
          const type_to_arrow_array_storage_t<Type>& arr, int64_t row) noexcept {
-  TENZIR_ASSERT(!arr.IsNull(row));
+  TENZIR_ASSERT_EXPENSIVE(!arr.IsNull(row));
   if constexpr (std::is_same_v<Type, null_type>) {
     return caf::none;
   } else if constexpr (detail::is_any_v<Type, bool_type, uint64_type,
@@ -149,13 +149,14 @@ value_at([[maybe_unused]] const Type& type,
   } else if constexpr (std::is_same_v<Type, int64_type>) {
     return int64_t{arr.GetView(row)};
   } else if constexpr (std::is_same_v<Type, duration_type>) {
-    TENZIR_ASSERT(
+    TENZIR_ASSERT_EXPENSIVE(
       caf::get<type_to_arrow_type_t<duration_type>>(*arr.type()).unit()
       == arrow::TimeUnit::NANO);
     return duration{arr.GetView(row)};
   } else if constexpr (std::is_same_v<Type, time_type>) {
-    TENZIR_ASSERT(caf::get<type_to_arrow_type_t<time_type>>(*arr.type()).unit()
-                  == arrow::TimeUnit::NANO);
+    TENZIR_ASSERT_EXPENSIVE(
+      caf::get<type_to_arrow_type_t<time_type>>(*arr.type()).unit()
+      == arrow::TimeUnit::NANO);
     return time{} + duration{arr.GetView(row)};
   } else if constexpr (std::is_same_v<Type, string_type>) {
     const auto str = arr.GetView(row);
@@ -164,12 +165,12 @@ value_at([[maybe_unused]] const Type& type,
     const auto str = arr.GetView(row);
     return {reinterpret_cast<const std::byte*>(str.data()), str.size()};
   } else if constexpr (std::is_same_v<Type, ip_type>) {
-    TENZIR_ASSERT(arr.byte_width() == 16);
+    TENZIR_ASSERT_EXPENSIVE(arr.byte_width() == 16);
     // NOLINTNEXTLINE(cppcoreguidelines-pro-bounds-pointer-arithmetic)
     const auto* bytes = arr.raw_values() + (row * 16);
     return ip::v6(std::span<const uint8_t, 16>{bytes, 16});
   } else if constexpr (std::is_same_v<Type, subnet_type>) {
-    TENZIR_ASSERT(arr.num_fields() == 2);
+    TENZIR_ASSERT_EXPENSIVE(arr.num_fields() == 2);
     auto network = value_at(
       ip_type{},
       *caf::get<type_to_arrow_array_t<ip_type>>(*arr.field(0)).storage(), row);
@@ -223,7 +224,7 @@ value_at([[maybe_unused]] const Type& type,
           // nop
         }
         value_type at(size_type i) const override {
-          TENZIR_ASSERT(!key_array->IsNull(value_offset + i));
+          TENZIR_ASSERT_EXPENSIVE(!key_array->IsNull(value_offset + i));
           if (item_array->IsNull(value_offset + i))
             return {value_at(key_type, *key_array, value_offset + i), {}};
           return {
@@ -280,8 +281,8 @@ template <concrete_type Type>
 view<type_to_data_t<Type>>
 value_at(const Type& type, const std::same_as<arrow::Array> auto& arr,
          int64_t row) noexcept {
-  TENZIR_ASSERT(type.to_arrow_type()->id() == arr.type_id());
-  TENZIR_ASSERT(!arr.IsNull(row));
+  TENZIR_ASSERT_EXPENSIVE(type.to_arrow_type()->id() == arr.type_id());
+  TENZIR_ASSERT_EXPENSIVE(!arr.IsNull(row));
   if constexpr (arrow::is_extension_type<type_to_arrow_type_t<Type>>::value)
     return value_at(type, *caf::get<type_to_arrow_array_t<Type>>(arr).storage(),
                     row);
@@ -291,7 +292,7 @@ value_at(const Type& type, const std::same_as<arrow::Array> auto& arr,
 
 data_view value_at(const type& type, const std::same_as<arrow::Array> auto& arr,
                    int64_t row) noexcept {
-  TENZIR_ASSERT(type.to_arrow_type()->id() == arr.type_id());
+  TENZIR_ASSERT_EXPENSIVE(type.to_arrow_type()->id() == arr.type_id());
   if (arr.IsNull(row))
     return caf::none;
   auto f = [&]<concrete_type Type>(const Type& type) noexcept -> data_view {