This PR makes failing assertions throw exceptions instead of aborting the process. These exceptions should generally not be caught and propagate all the way to our actor framework. Actors provide sufficient isolation such that the system can go back to a healthy state by just terminating the affected actor. Furthermore, this PR enables `TENZIR_ASSERT` by default in release builds. It also makes it so that `TENZIR_ASSERT(...)` can be used with format strings. The macros `TENZIR_UNREACHABLE` and `TENZIR_TODO` were changed to directly throw, without going through the assertion macros. Some places used `TENZIR_ASSERT` with expensive-to-evaluate expressions. Those were changed to `TENZIR_ASSERT_EXPENSIVE`, which continues to be disabled by default.

### Benchmark

To prevent performance regressions, I benchmarked this PR against main. 

~~~bash
#!/usr/bin/env bash

set -euo pipefail
rm -r tenzir.db
trap 'trap " " SIGTERM; kill 0; wait' SIGINT SIGTERM
coproc NODE { exec tenzir-node -q --print-endpoint; }
read -r -u "${NODE[0]}" TENZIR_ENDPOINT
IFS=':' read -r -a hostname_and_port <<< "${TENZIR_ENDPOINT}"
while ! lsof -i ":${hostname_and_port[1]}"; do
  echo "Waiting for node to be reachable..."
  sleep 1
done

echo "Spawning M57 Suricata pipeline"
tenzir -q "from suricata.json.zst read suricata --no-infer | where #schema != \"suricata.stats\" | import"

echo "Spawning M57 Zeek pipeline"
tenzir -q 'from zeek-all.log.zst read zeek-tsv | import'
tenzir -q 'from zeek-all.log.zst read zeek-tsv | import'
tenzir -q 'from zeek-all.log.zst read zeek-tsv | import'

echo "Rebuilding demo data"
tenzir-ctl -q rebuild --undersized --all --parallel=4

echo "Counting"
tenzir -q "export | summarize count(.)"

echo "Exporting"
tenzir -q "export | where :ip != 31.48.194.193 | head 500000 | where :ip != 54.193.192.5 | write json -c" > /dev/null

# NOTE: This benchmark exposed that waiting for proper node termination can sometimes deadlock. 
# This also happens on `main`. Hence, this benchmark directly kills the process. The termination 
# issue needs to be addressed separately.
echo "Done"
# kill "$NODE_PID"
# wait "$NODE_PID"
kill -9 "$NODE_PID"
~~~

### Results

~~~
$ hyperfine -- ./bench.bash
  # --- before (main) --- 
  Time (mean ± σ):     24.597 s ±  0.285 s    [User: 37.853 s, System: 3.196 s]
  Range (min … max):   24.211 s … 25.022 s    10 runs
  # --- after (this PR) ---
  Time (mean ± σ):     25.260 s ±  0.290 s    [User: 38.024 s, System: 3.133 s]
  Range (min … max):   24.804 s … 25.811 s    10 runs
  # --- all assertions disabled (not this PR) ---
  Time (mean ± σ):     24.444 s ±  0.184 s    [User: 37.602 s, System: 3.096 s]
  Range (min … max):   24.189 s … 24.745 s    10 runs
~~~